adjustments to the pieces ws manages and speed improvments
