+++ api/__init__.py
@@ -1,0 +1,1 @@
+from .pieces_ws import WebSocketManager
